+++ Percusion.java
@@ -2,7 +2,7 @@ public class Percusion extends Producto {
     String elementoPercutor; 
     String elementoVibrante;
 
-    public Percusion(String id, double precio, String mar, String mo, String n, String mat, int p, String percutor, String vibrante){
+    public Percusion(String id, double precio, String mar, String mo, String n, String mat, double p, String percutor, String vibrante){
         super(id, precio, mar, mo, n, mat, p); 
         this.elementoPercutor = percutor; 
         this.elementoVibrante = vibrante;
